High Order Component withGestureList add the handle gesture functionality, and Animated styles for gestures
